requirements.txt
django
gunicorn
requests
mozilla-django-oidc
djangorestframework
